takeoff/requirements.txt
anthropic
python-dotenv
Pillow
